Movement DAO - Asset Upload 2020 (#2573)
* The Movement DAO - Asset Upload 2020

Asset upload #1. 
The Movement is a fully decentralized organization for borderless collaboration. Driven by the vision of building bridges across borders and cultures, it is our most fundamental aim to facilitate the creation of value for the benefit of humanity. Dedicated to bringing about new opportunities for all, the organization is collectively created and owned by all who take part in its ongoing development and governance.

* Movement DAO Asset Upload

The Movement is a fully decentralized organization for borderless collaboration. Driven by the vision of building bridges across borders and cultures, it is our most fundamental aim to facilitate the creation of value for the benefit of humanity. Dedicated to bringing about new opportunities for all, the organization is collectively created and owned by all who take part in its ongoing development and governance.

* update pr/2573

delete wrong folder
checksum folder 0x3D46454212c61ECb7b31248047Fa033120B88668

Co-authored-by: zachzwei <35627271+zachzwei@users.noreply.github.com>